- Revamped upload system to work with dash_upload library. Uploads are now copied to a temporary folder, read, and then deleted. This should allow for processing of larger datasets (ESRF)
- EDX compiler updated to work with the new method
- Dektak compiler updated to work with the new method
- Moke compiler updated to work with the new method
