+++ .gitignore
@@ -19,3 +19,4 @@ _autosummary/
 venv
 data/
 results/
+/uploads/
