+++ modules/callbacks/callbacks_edx.py
@@ -26,9 +26,14 @@ def callbacks_edx(app):
                    Output("edx_heatmap_select", "value")],
                   Input("hdf5_path_store", "data"))
     def edx_update_element_list(hdf5_path):
+        hdf5_path = Path(hdf5_path)
+
         if hdf5_path is None:
             raise PreventUpdate
-        edx_element_list = get_quantified_elements(hdf5_path)
+
+        with h5py.File(hdf5_path, 'r') as hdf5_file:
+            edx_group = hdf5_file['edx']
+            edx_element_list = get_quantified_elements(edx_group)
         return edx_element_list, edx_element_list[0]
 
     # Callback to plot EDX heatmap
@@ -44,7 +49,6 @@ def callbacks_edx(app):
     )
     def edx_update_heatmap(heatmap_select, z_min, z_max, precision, hdf5_path):
         hdf5_path = Path(hdf5_path)
-
         if hdf5_path is None:
             raise PreventUpdate
 
@@ -52,7 +56,10 @@ def callbacks_edx(app):
             z_min = None
             z_max = None
 
-        edx_df = edx_make_results_dataframe_from_hdf5(hdf5_path)
+        with h5py.File(hdf5_path, 'r') as hdf5_file:
+            edx_group = hdf5_file['edx']
+            edx_df = edx_make_results_dataframe_from_hdf5(edx_group)
+
         fig = make_heatmap_from_dataframe(edx_df, values=heatmap_select, z_min=z_min, z_max=z_max, precision=precision)
 
         z_min = np.round(fig.data[0].zmin, precision)
@@ -73,17 +80,19 @@ def callbacks_edx(app):
         Input("edx_position_store", "data"),
     )
     def edx_update_plot(hdf5_path, position):
+        hdf5_path = Path(hdf5_path)
         if hdf5_path is None:
             raise PreventUpdate
         if position is None:
             raise PreventUpdate
 
-        hdf5_path = Path(hdf5_path)
-
         target_x = position[0]
         target_y = position[1]
 
-        measurement_df = edx_get_measurement_from_hdf5(hdf5_path, target_x, target_y)
+        with h5py.File(hdf5_path, 'r') as hdf5_file:
+            edx_group = hdf5_file['edx']
+            measurement_df = edx_get_measurement_from_hdf5(edx_group, target_x, target_y)
+
         fig = edx_plot_measurement_from_dataframe(measurement_df)
 
         return fig
