+++ modules/hdf5_compilers/hdf5compile_profil.py
@@ -6,9 +6,12 @@ from ..hdf5_compilers.hdf5compile_base import *
 import io
 
 
-def read_header_from_dektak(file_string):
+def read_header_from_dektak(file_path):
     header_dict = {}
-    lines = io.StringIO(file_string).readlines()
+
+    with open(file_path, 'r') as file:
+        lines = file.readlines()
+
     for line in lines[4:45]:
         split = line.strip().split(',')
         key, value = split[0], split[-1]
@@ -18,9 +21,8 @@ def read_header_from_dektak(file_string):
     del header_dict['FullFilename']
     return header_dict
 
-def read_data_from_dektak(file_string, header_length=46):
-    file = io.StringIO(file_string)
-    asc2d_dataframe = pd.read_csv(file, skiprows=header_length)
+def read_data_from_dektak(file_path, header_length=46):
+    asc2d_dataframe = pd.read_csv(file_path, skiprows=header_length)
     asc2d_dataframe.rename(columns={' z(raw/unitless)': 'profile'}, inplace=True)
     asc2d_dataframe.rename(columns={'y(um)': 'distance'}, inplace=True)
     return asc2d_dataframe
@@ -37,7 +39,7 @@ def set_instrument_from_dict(header_dict, node):
     Writes the contents of the moke_dict dictionary to the HDF5 node.
 
     Args:
-        moke_dict (dict): A dictionary containing the MOKE data and metadata, generated by the read_header_from_moke and read_data_from_moke functions.
+        header_dict (dict): A dictionary containing the MOKE data and metadata, generated by the read_header_from_moke and read_data_from_moke functions.
         node (h5py.Group): The HDF5 group to write the data to.
     Returns:
         None
@@ -51,38 +53,44 @@ def set_instrument_from_dict(header_dict, node):
     return None
 
 
-def write_dektak_to_hdf5(hdf5_path, measurement_dict, mode='a'):
+def write_dektak_to_hdf5(hdf5_path, source_path, mode='a'):
+    if isinstance (hdf5_path, str):
+        hdf5_path = Path(hdf5_path)
+    if isinstance(source_path, str):
+        source_path = Path(source_path)
+
     with h5py.File(hdf5_path, mode) as hdf5_file:
-        for file_name, file_string in measurement_dict.items():
-            if file_name.endswith('.asc2d'):
-                header_dict = read_header_from_dektak(file_string)
-                asc2d_dataframe = read_data_from_dektak(file_string)
-                scan_number = header_dict['TargetName']
+        for file_name in source_path.rglob('*.asc2d'):
+            file_path = source_path / file_name
 
-                x_pos, y_pos = position_from_tuple(scan_number)
+            header_dict = read_header_from_dektak(file_path)
+            asc2d_dataframe = read_data_from_dektak(file_path)
+            scan_number = header_dict['TargetName']
 
-                scan_group = f"/profil/scan_{scan_number}"
-                scan = hdf5_file.create_group(scan_group)
+            x_pos, y_pos = position_from_tuple(scan_number)
 
-                # Instrument group for metadata
-                instrument = scan.create_group("instrument")
-                instrument.attrs["NX_class"] = "HTinstrument"
-                instrument["x_pos"] = convertFloat(x_pos)
-                instrument["y_pos"] = convertFloat(y_pos)
-                instrument["x_pos"].attrs["units"] = "mm"
-                instrument["y_pos"].attrs["units"] = "mm"
+            scan_group = f"/profil/scan_{scan_number}"
+            scan = hdf5_file.create_group(scan_group)
 
-                set_instrument_from_dict(header_dict, instrument)
+            # Instrument group for metadata
+            instrument = scan.create_group("instrument")
+            instrument.attrs["NX_class"] = "HTinstrument"
+            instrument["x_pos"] = convertFloat(x_pos)
+            instrument["y_pos"] = convertFloat(y_pos)
+            instrument["x_pos"].attrs["units"] = "mm"
+            instrument["y_pos"].attrs["units"] = "mm"
 
-                # Measurement group for data
-                data = scan.create_group("measurement")
-                data.attrs["NX_class"] = "HTmeasurement"
-                for col in asc2d_dataframe.columns:
-                    node = data.create_dataset(col, data=np.array(asc2d_dataframe[col]), dtype='float')
-                    if col == 'profile':
-                        node.attrs['unit'] = 'nm'
-                    elif col == 'distance':
-                        node.attrs['unit'] = 'μm'
+            set_instrument_from_dict(header_dict, instrument)
+
+            # Measurement group for data
+            data = scan.create_group("measurement")
+            data.attrs["NX_class"] = "HTmeasurement"
+            for col in asc2d_dataframe.columns:
+                node = data.create_dataset(col, data=np.array(asc2d_dataframe[col]), dtype='float')
+                if col == 'profile':
+                    node.attrs['unit'] = 'nm'
+                elif col == 'distance':
+                    node.attrs['unit'] = 'μm'
 
     return None
