+++ requirements.txt
@@ -16,3 +16,4 @@ dash-bootstrap-components~=1.7.1
 h5py~=3.12.1
 scikit-learn~=1.6.1
 stringcase~=1.2.0
+dash-uploader
